requirements.txt
Django==5.1.4
openpyxl==3.1.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
